+++ codegen/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "v_escape_codegen"
+version = "0.1.0"
+authors = ["Juan Aguilar Santillana <mhpoin@gmail.com>"]
+description = "Code generator package for v_escape"
+edition = "2021"
+license = "MIT/Apache-2.0"
+readme = "../README.md"
+repository = "https://github.com/botika/v_escape"
+workspace = ".."
+
+[dependencies]
+syn = "1.0"
+quote = "1.0"
+proc-macro2 = "^1.0.37"
+clap = { version = "3.1", default-features = false, features = ["derive", "std"] }
+toml = "0.5"
+serde = { version = "1.0", features = ["derive"] }
