+++ codegen/src/chars.rs
