+++ codegen/src/ranges/mod.rs
@@ -1,51 +1,84 @@
-#[macro_use]
+use proc_macro2::{Span, TokenStream};
+use quote::quote;
+use syn::Ident;
+
 mod avx;
-#[macro_use]
-mod sse;
-#[macro_use]
 mod switch;
 
-#[macro_export]
-#[doc(hidden)]
-macro_rules! escape_ranges {
-    (avx2 $($t:tt)+) => {
+use self::switch::Switch;
+
+#[derive(Copy, Clone)]
+enum Feature {
+    Avx2,
+    Sse2,
+}
+use Feature::*;
+
+fn to_str(f: Feature) -> &'static str {
+    match f {
+        Avx2 => "avx2",
+        Sse2 => "sse2",
+    }
+}
+
+pub struct ArgLoop {
+    len: Ident,
+    ptr: Ident,
+    start_ptr: Ident,
+    end_ptr: Ident,
+    s: Switch,
+}
+
+fn to_loop(f: Feature, arg: ArgLoop) -> TokenStream {
+    match f {
+        Avx2 => avx::loop_range_switch_avx2(arg),
+        Sse2 => quote! {},
+    }
+}
+
+fn ident(s: &str) -> Ident {
+    Ident::new(s, Span::call_site())
+}
+
+fn escape_range(s: switch::Switch, f: Feature) -> TokenStream {
+    let feature = to_str(f);
+    let len = ident("len");
+    let start_ptr = ident("start_ptr");
+    let end_ptr = ident("end_ptr");
+    let ptr = ident("ptr");
+    let loops = to_loop(
+        f,
+        ArgLoop {
+            len: len.clone(),
+            ptr: ptr.clone(),
+            end_ptr: end_ptr.clone(),
+            start_ptr: start_ptr.clone(),
+            s,
+        },
+    );
+
+    quote! {
         #[inline]
-        #[target_feature(enable = "avx2")]
-        $crate::escape_ranges!(impl $crate::loop_range_switch_avx2 where $($t)+);
-    };
-    (sse2 $($t:tt)+) => {
-        #[inline]
-        #[target_feature(enable = "sse2")]
-        $crate::escape_ranges!(impl $crate::loop_range_switch_sse2 where $($t)+);
-    };
-    (impl $loops:path where ($T:ident, $Q:ident, $Q_LEN:ident) $($t:tt)+) => {
+        #[target_feature(enable = #feature)]
         pub unsafe fn escape(bytes: &[u8], fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
 
-            let len = bytes.len();
-            let start_ptr = bytes.as_ptr();
-            let end_ptr = bytes[len..].as_ptr();
-            let mut ptr = start_ptr;
+            let #len = bytes.len();
+            let #start_ptr = bytes.as_ptr();
+            let #end_ptr = bytes[len..].as_ptr();
+            let mut #ptr = #start_ptr;
 
             let mut start = 0;
 
             macro_rules! mask_bodies_callback {
                 ($callback:path) => {
-                    // Format bytes in the mask that starts in the current pointer
                     macro_rules! mask_bodies {
                         ($mask:ident, $at:ident, $cur:ident, $ptr:ident) => {
-                            // Calls macro `bodies!` at position `$at + $cur`
-                            // of byte `*$ptr` + `$curr` with macro `$crate::mask_body!`
                             $callback!($T, $Q, $Q_LEN, $at + $cur, *$ptr.add($cur), start, fmt, bytes, $crate::mask_body);
 
-                            // Create binary vector of all zeros except
-                            // position `$curr` and xor operation with `$mask`
                             $mask ^= 1 << $cur;
-                            // Test vs Check  if `$mask` is empty
                             if $mask == 0 {
                                 break;
                             }
-
-                            // Get to the next possible escape character avoiding zeros
                             $cur = $mask.trailing_zeros() as usize;
                         };
                     }
@@ -54,17 +87,12 @@ macro_rules! escape_ranges {
 
             $crate::mask_bodies_escaping!($($t)+);
 
-            // Macro to write with mask
             macro_rules! write_mask {
                 ($mask:ident, $ptr:ident) => {{
-                    // Reference to the start of mask
-                    let at = $crate::sub!($ptr, start_ptr);
-                    // Get to the first possible escape character avoiding zeros
+                    let at = $crate::sub!($ptr, #start_ptr);
                     let mut cur = $mask.trailing_zeros() as usize;
 
                     loop {
-                        // Writing in `$fmt` with `$mask`
-                        // The main loop will break when mask == 0
                         mask_bodies!($mask, at, cur, $ptr);
                     }
 
@@ -72,17 +100,16 @@ macro_rules! escape_ranges {
                 }};
             }
 
-            // Write a sliced mask
             macro_rules! write_forward {
                 ($mask: ident, $align:ident) => {{
-                    let at = $crate::sub!(ptr, start_ptr);
+                    let at = $crate::sub!(#ptr, #start_ptr);
                     let mut cur = $mask.trailing_zeros() as usize;
 
                     while cur < $align {
                         mask_bodies!($mask, at, cur, ptr);
                     }
 
-                    debug_assert_eq!(at, $crate::sub!(ptr, start_ptr))
+                    debug_assert_eq!(at, $crate::sub!(#ptr, #start_ptr))
                 }};
             }
 
@@ -95,7 +122,7 @@ macro_rules! escape_ranges {
                                     $T,
                                     $Q,
                                     $Q_LEN,
-                                    $crate::sub!(ptr, start_ptr),
+                                    $crate::sub!(#ptr, #start_ptr),
                                     *ptr,
                                     start,
                                     fmt,
@@ -116,8 +143,8 @@ macro_rules! escape_ranges {
                                         $T,
                                         $Q,
                                         $Q_LEN,
-                                        $crate::sub!(ptr, start_ptr),
-                                        *ptr,
+                                        $crate::sub!(#ptr, #start_ptr),
+                                        *#ptr,
                                         start,
                                         fmt,
                                         bytes,
@@ -133,162 +160,15 @@ macro_rules! escape_ranges {
 
             $crate::fallback_escaping!($($t)+);
 
-            $loops!((len, ptr, start_ptr, end_ptr) $($t)+);
-
-            // Write since start to the end of the slice
-            debug_assert!(start <= len);
-            if start < len {
-                fmt.write_str(std::str::from_utf8_unchecked(&bytes[start..len]))?;
+            #loops
+            debug_assert!(start <= #len);
+            if start < #len {
+                fmt.write_str(std::str::from_utf8_unchecked(&bytes[start..#len]))?;
             }
 
             Ok(())
         }
-    };
-}
-
-#[macro_export]
-#[doc(hidden)]
-macro_rules! escape_ranges_ptr {
-    (avx2 $($t:tt)+) => {
-        #[inline]
-        #[target_feature(enable = "avx2")]
-        $crate::escape_ranges_ptr!(impl $crate::loop_range_switch_avx2 where $($t)+);
-    };
-    (sse2 $($t:tt)+) => {
-        #[inline]
-        #[target_feature(enable = "sse2")]
-        $crate::escape_ranges_ptr!(impl $crate::loop_range_switch_sse2 where $($t)+);
-    };
-    (impl $loops:path where ($T:ident, $Q:ident, $Q_LEN:ident) $($t:tt)+) => {
-        pub unsafe fn f_escape(bytes: &[u8], buf: &mut [std::mem::MaybeUninit<u8>]) -> Option<usize> {
-            let mut buf_cur = 0;
-
-            let len = bytes.len();
-            let start_ptr = bytes.as_ptr();
-            let end_ptr = bytes[len..].as_ptr();
-            let mut ptr = start_ptr;
-
-            let mut start = 0;
-
-            macro_rules! mask_bodies_callback {
-                ($callback:path) => {
-                    // Format bytes in the mask that starts in the current pointer
-                    macro_rules! mask_bodies {
-                        ($mask:ident, $at:ident, $cur:ident, $ptr:ident) => {
-                            // Calls macro `bodies!` at position `$at + $cur`
-                            // of byte `*$ptr` + `$curr` with macro `$crate::mask_body!`
-                            $callback!($T, $Q, $Q_LEN, $at + $cur, *$ptr.add($cur), start, buf_cur, buf, start_ptr, $crate::mask_body_ptr);
-
-                            // Create binary vector of all zeros except
-                            // position `$curr` and xor operation with `$mask`
-                            $mask ^= 1 << $cur;
-                            // Test vs Check  if `$mask` is empty
-                            if $mask == 0 {
-                                break;
-                            }
-
-                            // Get to the next possible escape character avoiding zeros
-                            $cur = $mask.trailing_zeros() as usize;
-                        };
-                    }
-                };
-            }
-
-            $crate::mask_bodies_escaping_ptr!($($t)+);
-
-            // Macro to write with mask
-            macro_rules! write_mask {
-                ($mask:ident, $ptr:ident) => {{
-                    // Reference to the start of mask
-                    let at = $crate::sub!($ptr, start_ptr);
-                    // Get to the first possible escape character avoiding zeros
-                    let mut cur = $mask.trailing_zeros() as usize;
-
-                    loop {
-                        // Writing in `$fmt` with `$mask`
-                        // The main loop will break when mask == 0
-                        mask_bodies!($mask, at, cur, $ptr);
-                    }
-
-                    debug_assert_eq!(at, $crate::sub!($ptr, start_ptr))
-                }};
-            }
-
-            // Write a sliced mask
-            macro_rules! write_forward {
-                ($mask: ident, $align:ident) => {{
-                    let at = $crate::sub!(ptr, start_ptr);
-                    let mut cur = $mask.trailing_zeros() as usize;
-
-                    while cur < $align {
-                        mask_bodies!($mask, at, cur, ptr);
-                    }
-
-                    debug_assert_eq!(at, $crate::sub!(ptr, start_ptr))
-                }};
-            }
-
-            macro_rules! fallback_callback {
-                (default) => {
-                    macro_rules! fallback {
-                        () => {
-                            while ptr < end_ptr {
-                                $crate::bodies_ptr!(
-                                    $T,
-                                    $Q,
-                                    $Q_LEN,
-                                    $crate::sub!(ptr, start_ptr),
-                                    *ptr,
-                                    start,
-                                    buf_cur,
-                                    buf,
-                                    start_ptr,
-                                    $crate::mask_body_ptr
-                                );
-                                ptr = ptr.offset(1);
-                            }
-                        };
-                    }
-                };
-                (one) => {
-                    macro_rules! fallback {
-                        () => {
-                            while ptr < end_ptr {
-                                if *ptr == $T {
-                                    $crate::bodies_exact_one_ptr!(
-                                        $T,
-                                        $Q,
-                                        $Q_LEN,
-                                        $crate::sub!(ptr, start_ptr),
-                                        *ptr,
-                                        start,
-                                        buf_cur,
-                                        buf,
-                                        start_ptr,
-                                        $crate::mask_body_ptr
-                                    );
-                                }
-                                ptr = ptr.offset(1);
-                            }
-                        };
-                    }
-                };
-            }
-
-            $crate::fallback_escaping!($($t)+);
-
-            $loops!((len, ptr, start_ptr, end_ptr) $($t)+);
-
-            // Write since start to the end of the slice
-            debug_assert!(start <= len);
-            if start < len {
-                let len = len - start;
-                $crate::write_ptr!(buf_cur, buf, start_ptr.add(start), len);
-            }
-
-            Some(buf_cur)
-        }
-    };
+    }
 }
 
 #[macro_export]
@@ -422,7 +302,7 @@ macro_rules! escape_ranges_bytes {
 
             // Write since start to the end of the slice
             debug_assert!(start <= len);
-            if start < len {
+            if start < #len {
                 $crate::write_bytes!(&bytes[start..], buf);
             }
         }
