+++ testco/src/_lib.rs
@@ -1,0 +1,8 @@
+new!(
+    60 -> "&lt;",
+    62 -> "&gt;",
+    38 -> "&amp;",
+    34 -> "&quot;",
+    39 -> "&#x27;",
+    47 -> "&#x2f;"
+);
