--- v_escape_derive/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "v_escape_derive"
-version = "0.9.1"
-authors = ["Juan Aguilar Santillana <mhpoin@gmail.com>"]
-description = "Procedural macro package for v_escape"
-documentation = "https://docs.rs/v_escape_derive"
-edition = "2018"
-keywords = ["derive", "escaping", "simd"]
-license = "MIT/Apache-2.0"
-readme = "../README.md"
-repository = "https://github.com/botika/v_escape"
-workspace = ".."
-
-[badges]
-maintenance = { status = "actively-developed" }
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = "1.0"
