+++ v_htmlescape/src/lib.rs
@@ -7,18 +7,7 @@
 //! ```
 //!
 macro_rules! build {
-    ($($t:tt)*) => {
-        v_escape::new!(
-            HTMLEscape;
-            60 -> "&lt;",
-            62 -> "&gt;",
-            38 -> "&amp;",
-            34 -> "&quot;",
-            39 -> "&#x27;",
-            47 -> "&#x2f;";
-            $($t)*
-        );
-    };
+    ($($t:tt)*) => {};
 }
 
 /// Without simd optimizations
